fix(Chat): Correctly initialize socket connection for chat system (#65)
* fix(Chat): Correctly initialize socket connection for chat system

Previously the socket connection is initialize in the socket context.
Socket is not initialized at the conversation page and a null error
is shown.

This commit ensure that whenever the chat is loaded the socket
connection is completed first before emitting any services

* refactor(chat): rename redux actions and move socket event to a file

This patch rename redux actions by removing the suffix action in
the names. The socket events are also move to socket.js under constants
directory.

* refactor(Chat): Change redux and file names

This patch gets rids of poor naming of redux store variables and
file names. ChatRoom is renamed to ChatConversation. ChatList is
renamed to ChatRooms. Prefix such as get, set, add, insert, are
used in the redux action.

* refactor(chat): move socket config to context and rename socket functions

This patch moves the socket connection to the socket context. The socket
function names are renamed.

* refactor(socket): wrap socket provider over authenticated app in app.js