feat(Socket): add socket listeners and skeleton redux actions for chat offers (#75)
* feat(Socket Services): add socket listeners and skeleton redux actions for chat offers

This patch allow socket events to accept offer, create offer, and decline offer. The redux
action has also been created. TODOs are added to show what is suppose to happen when data
is received from socketEvent.

Note that the redux actions created do not do anything at the moment. The socket services
are setup with the correct parameters.

* docs(Socket): document Socket services with examples

This patch only adds comments to each socket services to demonstrate
usage. The comments are for both socketRequest and socketResponse.

Note that at the time of writing this documentation, the accept offer
service was found to be broken.

* fix(Access Token): Remove access token

This patch removes access token left by accident