feat(Offer): add functionality to accept offer (#78)
* feat(Offer): add functionality to accept offer

This patch includes the functionality for user to accept offer.
When the accept button is clicked the event is emitted to the
backend to indicate that the offer has been accepted. A socket
response is returned to indicate that the offer has successfully
been accepted. The buttons that show the offer in the conversation
will be hidden.

* style(Comments): Remove comments in ChatMessage

This patch remove the comments in ChatMessage. This comments are
just to help developer to know what the socket response is.

* fix(Offer): change relative imports to absolute imports

* fix(Offer): migrate redux store and remove disable param reassign

This patch migrate redux store to version 1 due to change in the
chat conversation model. The eslinting for param reassign is also
removed in ChatDux to prevent the code from being cluttered.