feat(Chat Messages): fix alignment issue chat messages using user type (#79)
* feat(Chat Messages): fix alignment issue chat messages using user type

This patch fixes the issue with the chat messages alignment. Previously,
the user type in messages and offer is sometimes null, resulting in the
misalignment of messages. This is not due to the backend issue. Rather,
the front end needs to pass in the user type to the socket service.

Currently, the messages are aligned using the user type. The sender and
receiver is identified by comparing the user type in the redux store.
Besides, the offers are also fixed with the user type. A photo is attached
to show how the offers look like after accepted for both seller and buyer.